# oapi_codegen/_oas/out/go/.docs/api/openapi.yaml
openapi: 3.0.2
info:
  title: Percona Everest schema
  version: 1.0.0
servers:
- url: /v1
...
paths:
  /kubernetes:
    get:
      description: List of registered kubernetes clusters
      operationId: listKubernetesClusters
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KubernetesCluster'
          description: Successful operation
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
          description: Unsuccessful operation
      summary: List of registered kubernetes clusters
      tags:
      - k8s
    post:
      description: Register Kubernetes cluster in the control plane
      operationId: registerKubernetes
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KubernetesCluster'
        description: Register a new kubernetes cluster in the control plane
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KubernetesCluster'
          description: Successful operation
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
          description: Unsuccessful operation
      summary: Register Kubernetes cluster in the control plane
      tags:
      - k8s
  /kubernetes/{kubernetesName}/database-cluster:
    get:
      description: List of created database clusters on provided kubernetes cluster
      operationId: listDatabases
      parameters:
      - description: Name of kubernetes Cluster
        explode: false
        in: path
        name: kubernetesName
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseClusterList'
          description: Successful operation
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
          description: Unsuccessful operation
      summary: List of created database clusters on provided kubernetes cluster
      tags:
      - k8s
    post:
      description: Create a database cluster on given kubernetes cluster
      operationId: createDatabaseCluster
      parameters:
      - description: Name of kubernetes Cluster
        explode: false
        in: path
        name: kubernetesName
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/database_cluster'
        description: Register a new kubernetes cluster in the control plane
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/database_cluster'
          description: Successful operation
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
          description: Unsuccessful operation
      summary: Create a database cluster on given kubernetes cluster
      tags:
      - k8s
components:
  schemas:
    BadRequest:
      description: Bad request response
      properties:
        message:
          type: string
      type: object
    KubernetesCluster:
      description: kubernetes object
      example:
        name: minikube
        kubeconfig: kubeconfig
      properties:
        name:
          example: minikube
          type: string
        kubeconfig:
          type: string
      type: object
...
    DatabaseClusterList:
      items:
        $ref: '#/components/schemas/database_cluster'
      type: array
    database_cluster:
      description: DatabaseCluster is the Schema for the databases API.
      example:
        metadata: "{}"
        apiVersion: apiVersion
        kind: kind
        spec:
          databaseType: databaseType
          dbInstance:
            diskSize: null
            memory: null
            storageClassName: storageClassName
            cpu: null
          backup:
            storages:
              key:
                volumeSpec:
                  emptyDir:
                    sizeLimit: null
                    medium: medium
                  persistentVolumeClaim:
                    storageClassName: storageClassName
                    dataSourceRef:
                      apiGroup: apiGroup
                      kind: kind
                      name: name
                      namespace: namespace
                    volumeName: volumeName
                    resources:
                      claims:
                      - name: name
                      - name: name
                      requests:
                        key: null
                      limits:
                        key: null
                    selector:
                      matchExpressions:
                      - values:
                        - values
                        - values
                        key: key
                        operator: operator
                      - values:
                        - values
                        - values
                        key: key
                        operator: operator
                      matchLabels:
                        key: matchLabels
                    accessModes:
                    - accessModes
                    - accessModes
                    dataSource:
                      apiGroup: apiGroup
                      kind: kind
                      name: name
                    volumeMode: volumeMode
                  hostPath:
                    path: path
                    type: type
                storageProvider:
                  bucket: bucket
                  storageClass: storageClass
                  containerName: containerName
                  credentialsSecret: credentialsSecret
                  endpointUrl: endpointUrl
                  prefix: prefix
                  region: region
                podSecurityContext:
                  runAsUser: 3
                  seLinuxOptions:
                    role: role
                    level: level
                    type: type
                    user: user
                  fsGroup: 7
                  seccompProfile:
                    localhostProfile: localhostProfile
                    type: type
                  windowsOptions:
                    gmsaCredentialSpec: gmsaCredentialSpec
                    runAsUserName: runAsUserName
                    hostProcess: true
                    gmsaCredentialSpecName: gmsaCredentialSpecName
                  fsGroupChangePolicy: fsGroupChangePolicy
                  supplementalGroups:
                  - 2
                  - 2
                  runAsGroup: 9
                  runAsNonRoot: true
                  sysctls:
                  - name: name
                    value: value
                  - name: name
                    value: value
                priorityClassName: priorityClassName
                verifyTLS: true
                annotations:
                  key: annotations
                resources:
                  claims:
                  - name: name
                  - name: name
                  requests:
                    key: null
                  limits:
                    key: null
                type: type
                labels:
                  key: labels
                nodeSelector:
                  key: nodeSelector
                runtimeClassName: runtimeClassName
                tolerations:
                - effect: effect
                  tolerationSeconds: 4
                  value: value
                  key: key
                  operator: operator
                - effect: effect
                  tolerationSeconds: 4
                  value: value
                  key: key
                  operator: operator
                containerSecurityContext:
                  privileged: true
                  runAsUser: 6
                  capabilities:
                    add:
                    - add
                    - add
                    drop:
                    - drop
                    - drop
                  seLinuxOptions:
                    role: role
                    level: level
                    type: type
                    user: user
                  seccompProfile:
                    localhostProfile: localhostProfile
                    type: type
                  windowsOptions:
                    gmsaCredentialSpec: gmsaCredentialSpec
                    runAsUserName: runAsUserName
                    hostProcess: true
                    gmsaCredentialSpecName: gmsaCredentialSpecName
                  procMount: procMount
                  allowPrivilegeEscalation: true
                  runAsGroup: 0
                  runAsNonRoot: true
                  readOnlyRootFilesystem: true
                schedulerName: schedulerName
                affinity:
                  nodeAffinity:
                    requiredDuringSchedulingIgnoredDuringExecution:
                      nodeSelectorTerms:
                      - matchExpressions:
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                        matchFields:
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                      - matchExpressions:
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                        matchFields:
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                    preferredDuringSchedulingIgnoredDuringExecution:
                    - preference:
                        matchExpressions:
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                        matchFields:
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                      weight: 5
                    - preference:
                        matchExpressions:
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                        matchFields:
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                      weight: 5
                  podAffinity:
                    requiredDuringSchedulingIgnoredDuringExecution:
                    - labelSelector:
                        matchExpressions:
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                        matchLabels:
                          key: matchLabels
                      namespaceSelector:
                        matchExpressions:
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                        matchLabels:
                          key: matchLabels
                      topologyKey: topologyKey
                      namespaces:
                      - namespaces
                      - namespaces
                    - labelSelector:
                        matchExpressions:
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                        matchLabels:
                          key: matchLabels
                      namespaceSelector:
                        matchExpressions:
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                        matchLabels:
                          key: matchLabels
                      topologyKey: topologyKey
                      namespaces:
                      - namespaces
                      - namespaces
                    preferredDuringSchedulingIgnoredDuringExecution:
                    - podAffinityTerm:
                        labelSelector:
                          matchExpressions:
                          - values:
                            - values
                            - values
                            key: key
                            operator: operator
                          - values:
                            - values
                            - values
                            key: key
                            operator: operator
                          matchLabels:
                            key: matchLabels
                        namespaceSelector:
                          matchExpressions:
                          - values:
                            - values
                            - values
                            key: key
                            operator: operator
                          - values:
                            - values
                            - values
                            key: key
                            operator: operator
                          matchLabels:
                            key: matchLabels
                        topologyKey: topologyKey
                        namespaces:
                        - namespaces
                        - namespaces
                      weight: 2
                    - podAffinityTerm:
                        labelSelector:
                          matchExpressions:
                          - values:
                            - values
                            - values
                            key: key
                            operator: operator
                          - values:
                            - values
                            - values
                            key: key
                            operator: operator
                          matchLabels:
                            key: matchLabels
                        namespaceSelector:
                          matchExpressions:
                          - values:
                            - values
                            - values
                            key: key
                            operator: operator
                          - values:
                            - values
                            - values
                            key: key
                            operator: operator
                          matchLabels:
                            key: matchLabels
                        topologyKey: topologyKey
                        namespaces:
                        - namespaces
                        - namespaces
                      weight: 2
                  podAntiAffinity:
                    requiredDuringSchedulingIgnoredDuringExecution:
                    - labelSelector:
                        matchExpressions:
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                        matchLabels:
                          key: matchLabels
                      namespaceSelector:
                        matchExpressions:
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                        matchLabels:
                          key: matchLabels
                      topologyKey: topologyKey
                      namespaces:
                      - namespaces
                      - namespaces
                    - labelSelector:
                        matchExpressions:
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                        matchLabels:
                          key: matchLabels
                      namespaceSelector:
                        matchExpressions:
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                        - values:
                          - values
                          - values
                          key: key
                          operator: operator
                        matchLabels:
                          key: matchLabels
                      topologyKey: topologyKey
                      namespaces:
                      - namespaces
                      - namespaces
                    preferredDuringSchedulingIgnoredDuringExecution:
                    - podAffinityTerm:
                        labelSelector:
                          matchExpressions:
                          - values:
                            - values
                            - values
                            key: key
                            operator: operator
                          - values:
                            - values
                            - values
                            key: key
                            operator: operator
                          matchLabels:
                            key: matchLabels
                        namespaceSelector:
                          matchExpressions:
                          - values:
                            - values
                            - values
                            key: key
                            operator: operator
                          - values:
                            - values
                            - values
                            key: key
                            operator: operator
                          matchLabels:
                            key: matchLabels
                        topologyKey: topologyKey
                        namespaces:
                        - namespaces
                        - namespaces
                      weight: 2
                    - podAffinityTerm:
                        labelSelector:
                          matchExpressions:
                          - values:
                            - values
                            - values
                            key: key
                            operator: operator
                          - values:
                            - values
                            - values
                            key: key
                            operator: operator
                          matchLabels:
                            key: matchLabels
                        namespaceSelector:
                          matchExpressions:
                          - values:
                            - values
                            - values
                            key: key
                            operator: operator
                          - values:
                            - values
                            - values
                            key: key
                            operator: operator
                          matchLabels:
                            key: matchLabels
                        topologyKey: topologyKey
                        namespaces:
                        - namespaces
                        - namespaces
                      weight: 2
            image: image
            imagePullPolicy: imagePullPolicy
            initImage: initImage
            schedule:
            - schedule: schedule
              keep: 5
              name: name
              compressionType: compressionType
              compressionLevel: 1
              enabled: true
              storageName: storageName
            - schedule: schedule
              keep: 5
              name: name
              compressionType: compressionType
              compressionLevel: 1
              enabled: true
              storageName: storageName
            serviceAccountName: serviceAccountName
            imagePullSecrets:
            - name: name
            - name: name
            annotations:
              key: annotations
            resources:
              claims:
              - name: name
              - name: name
              requests:
                key: null
              limits:
                key: null
            containerSecurityContext:
              privileged: true
              runAsUser: 6
              capabilities:
                add:
                - add
                - add
                drop:
                - drop
                - drop
              seLinuxOptions:
                role: role
                level: level
                type: type
                user: user
              seccompProfile:
                localhostProfile: localhostProfile
                type: type
              windowsOptions:
                gmsaCredentialSpec: gmsaCredentialSpec
                runAsUserName: runAsUserName
                hostProcess: true
                gmsaCredentialSpecName: gmsaCredentialSpecName
              procMount: procMount
              allowPrivilegeEscalation: true
              runAsGroup: 0
              runAsNonRoot: true
              readOnlyRootFilesystem: true
            enabled: true
            labels:
              key: labels
          databaseImage: databaseImage
          databaseConfig: databaseConfig
          loadBalancer:
            image: image
            size: 1
            configuration: configuration
            exposeType: exposeType
            annotations:
              key: annotations
            loadBalancerSourceRanges:
            - loadBalancerSourceRanges
            - loadBalancerSourceRanges
            resources:
              claims:
              - name: name
              - name: name
              requests:
                key: null
              limits:
                key: null
            type: type
            trafficPolicy: trafficPolicy
          monitoring:
            imagePullPolicy: imagePullPolicy
            pmm:
              image: image
              password: password
              serverUser: serverUser
              publicAddress: publicAddress
              login: login
              serverHost: serverHost
            runtimeClassName: runtimeClassName
            resources:
              claims:
              - name: name
              - name: name
              requests:
                key: null
              limits:
                key: null
            containerSecurityContext:
              privileged: true
              runAsUser: 6
              capabilities:
                add:
                - add
                - add
                drop:
                - drop
                - drop
              seLinuxOptions:
                role: role
                level: level
                type: type
                user: user
              seccompProfile:
                localhostProfile: localhostProfile
                type: type
              windowsOptions:
                gmsaCredentialSpec: gmsaCredentialSpec
                runAsUserName: runAsUserName
                hostProcess: true
                gmsaCredentialSpecName: gmsaCredentialSpecName
              procMount: procMount
              allowPrivilegeEscalation: true
              runAsGroup: 0
              runAsNonRoot: true
              readOnlyRootFilesystem: true
          secretsName: secretsName
          pause: true
          clusterSize: 7
        status:
          size: 1
          ready: 1
          host: host
          message: message
          status: status
      properties:
        apiVersion:
          description: "APIVersion defines the versioned schema of this representation\
            \ of an object. Servers should convert recognized schemas to the latest\
            \ internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources"
          type: string
        kind:
          description: "Kind is a string value representing the REST resource this\
            \ object represents. Servers may infer this from the endpoint the client\
            \ submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds"
          type: string
        metadata:
          type: object
        spec:
          $ref: '#/components/schemas/database_cluster_spec'
        status:
          $ref: '#/components/schemas/database_cluster_status'
      type: object
...
    database_cluster_status:
      description: DatabaseClusterStatus defines the observed state of Database.
      example:
        size: 1
        ready: 1
        host: host
        message: message
        status: status
      properties:
        host:
          type: string
        message:
          type: string
        ready:
          format: int32
          type: integer
        size:
          format: int32
          type: integer
        status:
          description: AppState is used to represent cluster's state.
          type: string
      type: object